requirements.txt
flask
pycryptodome
jwt
requests
Django>=4.0
djangorestframework
PyJWT
python-dotenv
